Изменение : Обновление libdcadec git-v0.1.0-9-g6cc57db; Изменение : Обновление MediaInfo git-v0.7.80-5-gbb1ea54; Изменение : Обновление ZenLib git-v0.4.32-11-gf36cf81;
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@985 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/ExtLib/MediaInfo/MediaInfo/Audio/File_DolbyE.cpp
@@ -554,7 +554,7 @@ const char*  DolbyE_ChannelLayout_PerProgram(int8u ProgramConfiguration, int8u P
     }
 };
 
-extern const float32 Mpegv_frame_rate[16];
+extern const float64 Mpegv_frame_rate[16];
 
 const bool Mpegv_frame_rate_type[16]=
 {false, false, false, false, false, false, true, true, true, false, false, false, false, false, false, false};
